Redo the compile warning for license_info.c fix - use size_t as the storage field for the return from strlen
git-svn-id: https://cctools.svn.sourceforge.net/svnroot/cctools/liblicense/trunk@8601 aec4f576-4910-0410-beeb-ecb0427469e7

+++ liblicense/license_info.c
@@ -189,13 +189,13 @@ ll_get_version (const ll_uri_t u)
     _ll_get_first (ll_get_attribute (u, LL_ATTRIBUTE_URI_VERSION, false));
   if (version)
     {
-      int c;
+      size_t c;
       int divisions;
       int position;
       ll_version_t result;
 
       divisions = 1;
-      for (c = 0; c < (int) strlen (version); ++c)
+      for (c = 0; c < strlen (version); ++c)
         {
           if (version[c] == '.')
             {
